[api-rewrite] working /info and /health
